bin: separate run-installed and run-inplace scripts
When xgrab is not installed you can run it via the bin/run-xgrab
command (which is like run-mozilla.sh) while when it is installed
you run it via the bin/xgrab command.

The rationale of this change is that I don't want "." to be in
the sys.path when we run the installed xgrab (it might not be safe
to fetch modules from the current dir).